refactor: change project architecture
