--- App/app/src/main/java/com/example/pokejournal/activities/MainActivity.java
@@ -1,32 +1,0 @@
-package com.example.pokejournal.activities;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-import com.example.pokejournal.R;
-
-
-public class MainActivity extends AppCompatActivity {
-Handler handler;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(MainActivity.this, navbar.class);
-                startActivity(intent);
-                finish();
-            }
-        }, 3000);
-
-    }
-
-
-}
