+++ App/app/src/main/java/com/example/pokejournal/application/activities/MainActivity.java
@@ -1,0 +1,32 @@
+package com.example.pokejournal.application.activities;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+
+import com.example.pokejournal.R;
+
+
+public class MainActivity extends AppCompatActivity {
+Handler handler;
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        handler = new Handler();
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent = new Intent(MainActivity.this, navbar.class);
+                startActivity(intent);
+                finish();
+            }
+        }, 3000);
+
+    }
+
+
+}
