+++ App/app/src/main/java/com/example/pokejournal/application/activities/PokeWidget.java
@@ -1,0 +1,38 @@
+package com.example.pokejournal.application.activities;
+
+import android.app.PendingIntent;
+import android.appwidget.AppWidgetManager;
+import android.appwidget.AppWidgetProvider;
+import android.content.Context;
+import android.widget.RemoteViews;
+
+import com.example.pokejournal.R;
+
+public class PokeWidget extends AppWidgetProvider {
+
+    static void updateAppWidget(Context context, AppWidgetManager appWidgetManager,
+                                int appWidgetId) {
+
+        CharSequence widgetText = context.getString(R.string.appwidget_text);
+
+        RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.poke_widget);
+        views.setTextViewText(R.id.appwidget_text, widgetText);
+
+        appWidgetManager.updateAppWidget(appWidgetId, views);
+    }
+
+    @Override
+    public void onUpdate(Context context, AppWidgetManager appWidgetManager, int[] appWidgetIds) {
+        for (int appWidgetId : appWidgetIds) {
+            updateAppWidget(context, appWidgetManager, appWidgetId);
+        }
+    }
+
+    @Override
+    public void onEnabled(Context context) {
+    }
+
+    @Override
+    public void onDisabled(Context context) {
+    }
+}
