+++ App/app/src/main/java/com/example/pokejournal/application/activities/t_deta_bulba.java
@@ -1,0 +1,31 @@
+package com.example.pokejournal.application.activities;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageView;
+
+import com.example.pokejournal.R;
+
+
+public class t_deta_bulba extends AppCompatActivity {
+    private ImageView icMenu;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_tdeta_bulba);
+
+        icMenu = findViewById(R.id.img_menu);
+        icMenu.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent bulbaIntent = new Intent(t_deta_bulba.this, navbar.class);
+                startActivity(bulbaIntent);
+            }
+        });
+
+
+    }
+}
