+++ App/app/src/main/java/com/example/pokejournal/domain/entities/pokejournal/PokemonTeam.java
@@ -1,0 +1,2 @@
+package com.example.pokejournal.domain.entities.pokejournal;
+public class PokemonTeam { }
