+++ App/app/src/main/java/com/example/pokejournal/domain/entities/pokejournal/User.java
@@ -1,0 +1,5 @@
+package com.example.pokejournal.domain.entities.pokejournal;
+
+public class User {
+
+}
