+++ App/app/src/main/java/com/example/pokejournal/domain/exceptions/MalformedException.java
@@ -1,0 +1,7 @@
+package com.example.pokejournal.domain.exceptions;
+
+public class MalformedException extends Exception{
+    public MalformedException(String msg){
+        super(msg);
+    }
+}
