+++ App/app/src/main/java/com/example/pokejournal/domain/exceptions/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.pokejournal.domain.exceptions;
+
+public class NotFoundException extends Exception {
+    public NotFoundException(String msg){
+        super(msg);
+    }
+}
