+++ App/app/src/main/java/com/example/pokejournal/domain/usecases/pokejournal/UserUsecase.java
@@ -1,0 +1,8 @@
+package com.example.pokejournal.domain.usecases.pokejournal;
+
+import com.example.pokejournal.domain.entities.pokejournal.User;
+
+public interface UserUsecase {
+    User registerUser(String username, String email, String password);
+    User login(String email, String password);
+}
