+++ App/app/src/main/res/layout/activity_detalhe_pokemon.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.detalhePokemonActivity">
+    tools:context=".application.activities.detalhePokemonActivity">
 
     <ImageView
         android:id="@+id/fundo_pokemon"
