+++ App/app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:background="#FFFFFF"
     android:backgroundTint="#FFFFFF"
     android:padding="16dp"
-    tools:context=".activities.MainActivity">
+    tools:context=".application.activities.MainActivity">
 
 
     <ImageView
