+++ App/app/src/main/res/layout/activity_quiz.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.QuizActivity">
+    tools:context=".application.activities.QuizActivity">
 
     <TextView
         android:id="@+id/pokemon_description"
