+++ App/app/src/main/res/layout/activity_tdeta_bulba.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.t_deta_bulba">
+    tools:context=".application.activities.t_deta_bulba">
 
     <ImageView
         android:id="@+id/fundo_pokemon"
